refactor: move DO transport behind host proxy
Move owner routing, hint caching, retry policy, and invoke/connect framing into the binding-scoped host adapter.
Keep tenant code limited to canonical request encoding and public Durable Object operations.
Preserve owner fences, WebSocket routing semantics, admission bounds, and internal authentication.
