+++ docs/source-map.md
@@ -33,6 +33,7 @@ are outside this map unless they own runtime or deployable service behavior.
 | `runtime/dispatch.js` and `runtime/dispatch/*` | Fetch, scheduled, queue, workflow dispatch, workflow step facade, replay cache, and deterministic workflow JSON helpers. |
 | `runtime/load.js` and `runtime/load/*` | Bundle decode, module rewrite, env construction, wrapper generation, injected runtime source ownership, and hidden binding stripping. `code-budget.js`, `injection-sources.js`, and `wrapper-generate.js` also own generated AI facade inclusion for AI-only bundles. |
 | `runtime/bindings/` | Host-side binding adapters for KV, D1, R2, Durable Objects, Workflows, ASSETS, service, queue, and AI. `runtime/bindings/do.js` and `runtime/bindings/workflow.js` own binding-scoped identity/authenticated transports; `runtime/bindings/ai.js` owns AI admission, provider resolution, public-only forwarding, and protocol orchestration; `runtime/bindings/ai-capacity.js` owns process-local pool accounting, lease transfer, metrics, deadlines, and watchdog cleanup; `runtime/bindings/ai-provider.js` owns official destinations, provider request policy, credential-bearing headers, and bounded response headers; `runtime/bindings/ai-sse.js` owns incremental SSE framing and terminal-event classification; `runtime/bindings/ai-websocket.js` owns bounded frame forwarding, model rewriting, idle detection, and close propagation. |
+| `runtime/do-client.js`, `runtime/_wdl-do-scoped-request.js`, `runtime/bindings/do.js`, `runtime/_wdl-do-transport.js` | Tenant DO namespace/id/stub facade and its binding-scoped request envelope; host-only identity attachment, owner routing, invoke/connect framing, retries, and response stripping. |
 | `runtime/ai-client.js` | Tenant-realm AI facade source for `run()`, `models()`, OpenAI-compatible fetch, streaming, and WebSocket helpers. |
 | `runtime/bindings/workflow.js`, `runtime/workflows-client.js`, `runtime/dispatch/workflow-*.js`, `runtime/load/wrapper-generate.js` | Workflow binding-scoped identity enforcement and transport, facade materialization, dispatch facade, replay cache, and step semantics. |
 | `runtime/tail-worker.js` / `runtime/tail-forwarder.js` | Workerd tail capture plus activation-gated append path for `wdl tail`. |
