+++ docs/source-map.zh.md
@@ -30,6 +30,7 @@
 | `runtime/dispatch.js` 和 `runtime/dispatch/*` | Fetch、scheduled、queue、workflow dispatch、workflow step facade、replay cache 和 deterministic workflow JSON helpers。 |
 | `runtime/load.js` 和 `runtime/load/*` | Bundle decode、module rewrite、env construction、wrapper generation、runtime 注入源码 ownership 和 hidden binding stripping。`code-budget.js`、`injection-sources.js` 和 `wrapper-generate.js` 还负责让只声明 AI 的 bundle 生成并计入 AI facade。 |
 | `runtime/bindings/` | KV、D1、R2、Durable Objects、Workflows、ASSETS、service、queue 和 AI 的 host-side binding adapters。`runtime/bindings/do.js` 与 `runtime/bindings/workflow.js` 负责 binding-scoped identity/authenticated transport；`runtime/bindings/ai.js` 负责 AI admission、provider 解析、public-only forwarding 和协议编排；`runtime/bindings/ai-capacity.js` 负责进程内 pool 计数、lease transfer、metrics、deadline 和 watchdog cleanup；`runtime/bindings/ai-provider.js` 负责 official destination、provider request policy、携带 credential 的 header 和有界 response header；`runtime/bindings/ai-sse.js` 负责增量 SSE framing 和 terminal-event classification；`runtime/bindings/ai-websocket.js` 负责有界 frame forwarding、model rewrite、idle detection 和 close propagation。 |
+| `runtime/do-client.js`、`runtime/_wdl-do-scoped-request.js`、`runtime/bindings/do.js`、`runtime/_wdl-do-transport.js` | Tenant DO namespace/id/stub facade 及其 binding-scoped request envelope；host-only identity attachment、owner routing、invoke/connect framing、retry 和 response stripping。 |
 | `runtime/ai-client.js` | Tenant realm 的 AI facade source，负责 `run()`、`models()`、OpenAI-compatible fetch、streaming 和 WebSocket helper。 |
 | `runtime/bindings/workflow.js`、`runtime/workflows-client.js`、`runtime/dispatch/workflow-*.js`、`runtime/load/wrapper-generate.js` | Workflow binding-scoped identity enforcement 与 transport、facade materialization、dispatch facade、replay cache 和 step semantics。 |
 | `runtime/tail-worker.js` / `runtime/tail-forwarder.js` | Workerd tail capture 和 `wdl tail` 的 activation-gated append path。 |
