+++ runtime/_wdl-owner-endpoint.js
@@ -1,2 +1,2 @@
-// Local adapter for the injected DO transport's relative module name.
+// Local adapter for the host DO transport's relative module name.
 export { prototypeGetter, validOwnerEndpointForService } from "../shared/owner-endpoint.js";
