+++ runtime/do-client.js
@@ -1,29 +1,8 @@
-import {
-  DO_CONNECT_URL,
-  DO_INVOKE_URL,
-  connectHeaders,
-  doOwnerHintCacheKey,
-  dispatchDoConnectWithHintCache,
-  dispatchDoInvokeWithHintCache,
-  fetchInvokeInit,
-  isWebSocketUpgrade,
-  replayOwnerUnavailableForFetch,
-  rpcInvokeInit,
-  rpcResultFromResponse,
-  scopedDoRequest,
-} from "./_wdl-do-transport.js";
-import { createOwnerHintCache } from "./_wdl-owner-hint-cache.js";
+import { scopedDoRequest } from "./_wdl-do-transport.js";
 import { requestIdFromOptions } from "./_wdl-request-id.js";
 
-const ownerHintCache = createOwnerHintCache();
-const intrinsicObjectHasOwn = Object.hasOwn;
 const intrinsicReflectApply = Reflect.apply;
 const intrinsicStringIsWellFormed = String.prototype.isWellFormed;
-
-/** @param {object} object @param {PropertyKey} key */
-function objectHasOwn(object, key) {
-  return intrinsicReflectApply(intrinsicObjectHasOwn, undefined, [object, key]);
-}
 
 /** @param {string} value */
 function isWellFormedUnicodeString(value) {
@@ -42,35 +21,15 @@ function requireObjectIdString(value, method) {
 }
 
 /**
- * @typedef {{ fetch(url: string, init?: RequestInit): Promise<Response> }} DoBackend
- * @typedef {{
- *   ns?: string,
- *   worker?: string,
- *   version?: string,
- *   doStorageId?: string,
- *   binding?: string,
- *   className?: string,
- * }} DurableObjectBindingProps
  * @typedef {{
  *   fetch?(request: Request): unknown,
  *   rpcObject?(objectName: string, method: string, args: unknown[], requestId: string | null): unknown,
  * }} DurableObjectBindingProxy
  * @typedef {{
- *   backend?: DoBackend,
- *   ownerNetwork?: DoBackend,
  *   requestId?: string,
  *   requestIdProvider?: () => string | null,
  * }} DurableObjectNamespaceOptions
  */
-
-export function clearDoOwnerHintsForTest() {
-  ownerHintCache.clearForTest();
-}
-
-/** @param {number} maxEntries */
-export function setDoOwnerHintMaxEntriesForTest(maxEntries) {
-  ownerHintCache.setMaxEntriesForTest(maxEntries);
-}
 
 class DurableObjectId {
   /** @param {string} name */
@@ -113,30 +72,11 @@ class DurableObjectStub {
   }
 }
 
-/** @param {DurableObjectBindingProps | undefined} props */
-function requireBindingProps(props) {
-  if (!props ||
-      typeof props.ns !== "string" ||
-      typeof props.worker !== "string" ||
-      typeof props.version !== "string" ||
-      typeof props.doStorageId !== "string" ||
-      typeof props.className !== "string") {
-    throw new Error("Durable Object binding metadata is not configured");
-  }
-  return /** @type {import("./_wdl-do-transport.js").DoBindingProps} */ (props);
-}
-
 export class DurableObjectNamespace {
-  /** @type {DoBackend | undefined} */
-  #backend;
   /** @type {null | ((objectName: string, request: Request, requestId: string | null) => Promise<Response>)} */
   #bindingFetch = null;
   /** @type {null | ((objectName: string, method: string, args: unknown[], requestId: string | null) => Promise<unknown>)} */
   #bindingRpcObject = null;
-  /** @type {DoBackend | undefined} */
-  #ownerNetwork;
-  /** @type {DurableObjectBindingProps | undefined} */
-  #props;
   /** @type {unknown} */
   #requestIdOptions = null;
 
@@ -154,44 +94,23 @@ export class DurableObjectNamespace {
         return response;
       }
       : null;
-    this.#bindingRpcObject = typeof proxy.rpcObject === "function"
-      ? async (objectName, method, args, requestId) => await proxy.rpcObject?.(objectName, method, args, requestId)
+    const rpcObject = proxy.rpcObject;
+    this.#bindingRpcObject = typeof rpcObject === "function"
+      ? async (objectName, method, args, requestId) => await intrinsicReflectApply(
+        rpcObject,
+        proxy,
+        [objectName, method, args, requestId]
+      )
       : null;
   }
 
   /**
-   * @param {DurableObjectBindingProps | DurableObjectBindingProxy | null | undefined} binding
-   * @param {DurableObjectNamespaceOptions | string} [options]
+   * @param {DurableObjectBindingProxy} binding
+   * @param {DurableObjectNamespaceOptions} [options]
    */
   constructor(binding, options = {}) {
-    const isMetadataBinding = binding && typeof binding === "object" && (
-      objectHasOwn(binding, "ns") ||
-      objectHasOwn(binding, "worker") ||
-      objectHasOwn(binding, "version") ||
-      objectHasOwn(binding, "doStorageId") ||
-      objectHasOwn(binding, "binding") ||
-      objectHasOwn(binding, "className")
-    );
-    if (binding && !isMetadataBinding) {
-      this.#setBindingProxy(/** @type {DurableObjectBindingProxy} */ (binding));
-    } else {
-      const props = /** @type {DurableObjectBindingProps | null | undefined} */ (binding);
-      this.#props = {
-        ns: props?.ns,
-        worker: props?.worker,
-        version: props?.version,
-        doStorageId: props?.doStorageId,
-        binding: props?.binding,
-        className: props?.className,
-      };
-    }
-    if (typeof options === "string") {
-      this.#requestIdOptions = { requestId: options };
-    } else {
-      this.#backend = options?.backend;
-      this.#ownerNetwork = options?.ownerNetwork;
-      this.#requestIdOptions = options || null;
-    }
+    this.#setBindingProxy(binding);
+    this.#requestIdOptions = options || null;
   }
 
   #currentRequestId() {
@@ -204,36 +123,7 @@ export class DurableObjectNamespace {
     if (this.#bindingFetch) {
       return await this.#bindingFetch(objectName, request, requestId);
     }
-    // Explicit backend options are the low-level transport construction path;
-    // runtime materialization supplies binding-scoped methods above.
-    const backend = this.#backend;
-    if (!backend || typeof backend.fetch !== "function") {
-      throw new Error("Durable Object backend is not configured");
-    }
-    const props = requireBindingProps(this.#props);
-    if (isWebSocketUpgrade(request)) {
-      const init = {
-        method: "GET",
-        headers: connectHeaders(props, objectName, request, requestId),
-      };
-      return await this.#dispatchWithOwnerHint(
-        dispatchDoConnectWithHintCache,
-        DO_CONNECT_URL,
-        "/internal/do/connect",
-        init,
-        doOwnerHintCacheKey(props, objectName),
-        false
-      );
-    }
-    const init = await fetchInvokeInit(props, objectName, request, requestId);
-    return await this.#dispatchWithOwnerHint(
-      dispatchDoInvokeWithHintCache,
-      DO_INVOKE_URL,
-      "/internal/do/invoke",
-      init,
-      doOwnerHintCacheKey(props, objectName),
-      replayOwnerUnavailableForFetch(request)
-    );
+    throw new Error("Durable Object binding fetch is not configured");
   }
 
   /** @param {string} objectName @param {string} method @param {unknown[]} args */
@@ -242,53 +132,7 @@ export class DurableObjectNamespace {
     if (this.#bindingRpcObject) {
       return await this.#bindingRpcObject(objectName, method, args, requestId);
     }
-    const backend = this.#backend;
-    if (!backend || typeof backend.fetch !== "function") {
-      throw new Error("Durable Object backend is not configured");
-    }
-    const props = requireBindingProps(this.#props);
-    const init = rpcInvokeInit(props, objectName, method, args, requestId);
-    const response = await this.#dispatchWithOwnerHint(
-      dispatchDoInvokeWithHintCache,
-      DO_INVOKE_URL,
-      "/internal/do/invoke",
-      init,
-      doOwnerHintCacheKey(props, objectName),
-      false
-    );
-    return await rpcResultFromResponse(response);
-  }
-
-  /**
-   * @param {typeof dispatchDoInvokeWithHintCache | typeof dispatchDoConnectWithHintCache} dispatch
-   * @param {string} routerUrl
-   * @param {string} ownerPath
-   * @param {RequestInit} init
-   * @param {string} hintKey
-   * @param {boolean} replayOwnerUnavailable
-   */
-  async #dispatchWithOwnerHint(dispatch, routerUrl, ownerPath, init, hintKey, replayOwnerUnavailable) {
-    const backend = this.#backend;
-    const ownerNetwork = this.#ownerNetwork;
-    if (!backend || typeof backend.fetch !== "function") {
-      throw new Error("Durable Object backend is not configured");
-    }
-    /** @type {import("./_wdl-do-transport.js").DoFetch | null} */
-    const ownerFetch = typeof ownerNetwork?.fetch === "function"
-      ? (url, requestInit) => ownerNetwork.fetch(url, requestInit)
-      : null;
-    /** @type {import("./_wdl-do-transport.js").DoFetch} */
-    const routerFetch = (url, requestInit) => backend.fetch(url, requestInit);
-    return await dispatch({
-      routerFetch,
-      routerUrl,
-      ownerFetch,
-      ownerPath,
-      init,
-      cache: ownerHintCache,
-      hintKey,
-      replayOwnerUnavailable,
-    });
+    throw new Error("Durable Object binding RPC is not configured");
   }
 
   /** @param {string} name */
