+++ tests/helpers/load-do-protocol.js
@@ -12,7 +12,7 @@ const SHARED_NS_PATTERN_URL = repositoryFileUrl("shared/ns-pattern.js");
 const SHARED_RESPOND_URL = repositoryFileUrl("shared/respond.js");
 const SHARED_UTF8_URL = repositoryFileUrl("shared/utf8.js");
 const WORKER_CONTRACT_URL = repositoryFileUrl("shared/worker-contract.js");
-const DO_TRANSPORT_URL = repositoryFileUrl("runtime/_wdl-do-transport.js");
+const DO_SCOPED_REQUEST_URL = repositoryFileUrl("runtime/_wdl-do-scoped-request.js");
 const DO_WIRE_GRAMMAR_URL = repositoryFileUrl("do-runtime/protocol/wire-grammar.js");
 const DO_ERRORS_URL = repositoryModuleDataUrl("do-runtime/protocol/errors.js", [
   [/from "shared-respond";/g, `from ${JSON.stringify(SHARED_RESPOND_URL)};`],
@@ -36,7 +36,7 @@ export function doProtocolDataUrl() {
       "shared-ns-pattern": SHARED_NS_PATTERN_URL,
       "shared-utf8": SHARED_UTF8_URL,
       "shared-worker-contract": WORKER_CONTRACT_URL,
-      "runtime-do-transport": DO_TRANSPORT_URL,
+      "_wdl-do-scoped-request.js": DO_SCOPED_REQUEST_URL,
     }),
   ]);
 }
