+++ tsconfig.json
@@ -20,6 +20,7 @@
       "runtime-dispatch-workflow-*": ["./runtime/dispatch/workflow-*.js"],
       "runtime-load-*": ["./runtime/load/*.js"],
       "runtime-do-transport": ["./runtime/_wdl-do-transport.js"],
+      "_wdl-do-scoped-request.js": ["./runtime/_wdl-do-scoped-request.js"],
       "runtime-owner-hint-cache": ["./runtime/_wdl-owner-hint-cache.js"],
       "runtime-state": ["./runtime/runtime.js"],
       "runtime-*": ["./runtime/*.js"],
